moved footer location
